move asexpectedtype to extensions
